video_extrac/requirements.txt
flask
google-cloud-storage
